refactor: se movio el store a src/app/store y se arreglaron las rutas
