+++ src/pages/scholarship-gallery/components/ScholarshipDetailModal.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux';
 import { Dialog } from 'primereact/dialog';
 import type { SB_Scholarship, ScholarshipRequirement } from '../../../shared/models/scholarshipsModel';
-import type { RootState } from '../../../store/store';
+import type { RootState } from '../../../app/store/store';
 import { ApplyScholarshipButton } from './ApplyScholarshipButton';
 
 interface ScholarshipDetailModalProps {
